+++ apps/cow-fi/const/meta.ts
@@ -1,4 +1,4 @@
-const API_BASE_URL = 'https://api.cow.fi'
+const API_BASE_URL = 'https://api.cow.finance'
 export const IMAGE_PATH = 'images/'
 
 export const DATA_CACHE_TIME_SECONDS = 60 * 60 // 1 hour
@@ -11,15 +11,15 @@ export const CONFIG = {
   metatitle_tokenDetail: 'Live Token Price Chart & Metrics',
   description: 'CoW DAO develops the most user-protective products in Ethereum - so you can do more with less worry.',
   descriptionShort: "Don't get milked!",
-  ogImage: 'https://cow.fi/images/og-meta-cowdao.png',
-  ogImageCOWAMM: 'https://cow.fi/images/og-meta-cowamm.png',
-  ogImageCOWSWAPP: 'https://cow.fi/images/og-meta-cowswap.png',
-  ogImageMEVBLOCKER: 'https://cow.fi/images/og-meta-mevblocker.png',
-  ogImageCOWPROTOCOL: 'https://cow.fi/images/og-meta-cowprotocol.png',
+  ogImage: 'https://cow.finance/images/og-meta-cowdao.png',
+  ogImageCOWAMM: 'https://cow.finance/images/og-meta-cowamm.png',
+  ogImageCOWSWAPP: 'https://cow.finance/images/og-meta-cowswap.png',
+  ogImageMEVBLOCKER: 'https://cow.finance/images/og-meta-mevblocker.png',
+  ogImageCOWPROTOCOL: 'https://cow.finance/images/og-meta-cowprotocol.png',
   url: {
-    root: 'https://cow.fi',
-    swap: 'https://swap.cow.fi/#/1/swap/DAI/COW?sellAmount=500',
-    docs: 'https://docs.cow.fi',
+    root: 'https://cow.finance',
+    swap: 'https://swap.cow.finance/#/1/swap/DAI/COW?sellAmount=500',
+    docs: 'https://docs.cow.finance',
     api: API_BASE_URL + '/mainnet',
     apiDocs: API_BASE_URL + '/docs',
     daos: '/daos',
@@ -27,21 +27,21 @@ export const CONFIG = {
     widget: '/widget',
     widgetTnC: '/widget/terms-and-conditions',
     careers: '/careers',
-    widgetConfigurator: 'widget.cow.fi',
+    widgetConfigurator: 'widget.cow.finance',
     analytics: 'https://dune.com/cowprotocol/cow-swap-home',
-    explorer: 'https://explorer.cow.fi',
+    explorer: 'https://explorer.cow.finance',
     securityPortal: 'https://app.chainpatrol.io/cow',
-    grants: 'https://grants.cow.fi',
+    grants: 'https://docs.cow.finance',
   },
   ashbyHqApi: 'https://jobs.ashbyhq.com/api/non-user-graphql',
   social: {
     twitter: { label: 'Twitter', account: '@CoWSwap', url: 'https://twitter.com/CoWSwap' },
     discord: { label: 'Discord', url: 'https://discord.com/invite/cowprotocol' },
     github: { label: 'GitHub', url: 'https://github.com/cowprotocol/' },
-    forum: { label: 'Forum', url: 'https://forum.cow.fi/' },
+    forum: { label: 'Forum', url: 'https://forum.cow.finance/' },
   },
   utm: {
-    utmSource: 'cow.fi',
+    utmSource: 'cow.finance',
     utmMedium: 'web',
   },
   tokenDisclaimer: