+++ libs/common-utils/src/explorer.ts
@@ -5,14 +5,14 @@ import { isBarn, isDev, isLocal, isPr, isStaging } from './environments'
 function _getExplorerUrlByEnvironment(): Record<ChainId, string> {
   let baseUrl: string | undefined
   if (isLocal || isDev || isPr) {
-    baseUrl = process.env.REACT_APP_EXPLORER_URL_DEV || 'https://dev.explorer.cow.fi'
+    baseUrl = process.env.REACT_APP_EXPLORER_URL_DEV || 'https://dev.explorer.cow.finance'
   } else if (isStaging) {
-    baseUrl = process.env.REACT_APP_EXPLORER_URL_STAGING || 'https://staging.explorer.cow.fi'
+    baseUrl = process.env.REACT_APP_EXPLORER_URL_STAGING || 'https://staging.explorer.cow.finance'
   } else if (isBarn) {
-    baseUrl = process.env.REACT_APP_EXPLORER_URL_BARN || 'https://barn.explorer.cow.fi'
+    baseUrl = process.env.REACT_APP_EXPLORER_URL_BARN || 'https://barn.explorer.cow.finance'
   } else {
     // Production by default
-    baseUrl = process.env.REACT_APP_EXPLORER_URL_PROD || 'https://explorer.cow.fi'
+    baseUrl = process.env.REACT_APP_EXPLORER_URL_PROD || 'https://explorer.cow.finance'
   }
 
   return {